Combine age and charting utilities into one file.
