+++ src/Shared/ChartingUtilities.cs
@@ -2,6 +2,7 @@
 // Copyright © 2007-2019 Apex Resource Management Solutions Ltd. (ApexRMS). All rights reserved.
 
 using System;
+using System.IO;
 using System.Linq;
 using System.Data;
 using System.Text;
@@ -15,115 +16,11 @@ namespace SyncroSim.STSim
 {
     internal static class ChartingUtilities
     {
-        /// <summary>
-        /// Determines if the specified descriptor has an age reference
-        /// </summary>
-        /// <param name="descriptor"></param>
-        /// <returns></returns>
-        public static bool HasAgeReference(ChartDescriptor descriptor)
-        {
-            if (descriptor.IncludeDataFilter != null && descriptor.IncludeDataFilter.Contains("AgeClass"))
-            {
-                return true;
-            }
-
-            if (descriptor.DisaggregateFilter != null && descriptor.DisaggregateFilter.Contains("AgeClass"))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        /// Determinies if a descriptor in the specified collection has an age reference
-        /// </summary>
-        /// <param name="descriptors"></param>
-        /// <returns></returns>
-        /// <remarks></remarks>
-        public static bool HasAgeReference(ChartDescriptorCollection descriptors)
-        {
-            foreach (ChartDescriptor d in descriptors)
-            {
-                if (HasAgeReference(d))
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        /// Updates the age class column for the specified table
-        /// </summary>
-        /// <param name="store"></param>
-        /// <param name="dataSheet"></param>
-        /// <remarks></remarks>
-        public static void UpdateAgeClassColumn(DataStore store, DataSheet dataSheet)
-        {
-            IEnumerable<AgeDescriptor> e = AgeUtilities.GetAgeDescriptors(dataSheet.Project);
-
-            if (e == null)
-            {
-                return;
-            }
-
-            if (!AnyDataExists(store, dataSheet))
-            {
-                return;
-            }
-
-            if (!AgeDataExists(store, dataSheet))
-            {
-                return;
-            }
-
-            StringBuilder sb = new StringBuilder();
-            Debug.Assert(!(e.ElementAtOrDefault(e.Count() - 1).MaximumAge.HasValue));
-
-            if (e.Count() > 1)
-            {
-                sb.AppendFormat(CultureInfo.InvariantCulture, 
-                    "UPDATE [{0}] SET AgeClass = CASE", 
-                    dataSheet.Name);
-
-                for (int i = 0; i < e.Count(); i++)
-                {
-                    AgeDescriptor d = e.ElementAtOrDefault(i);
-
-                    if (d.MaximumAge.HasValue)
-                    {
-                        Debug.Assert(i < e.Count() - 1);
-
-                        sb.AppendFormat(CultureInfo.InvariantCulture, 
-                            " WHEN AgeMin >= {0} And AgeMax <= {1} THEN {2}", 
-                            d.MinimumAge, d.MaximumAge.Value, d.MinimumAge);
-                    }
-                    else
-                    {
-                        Debug.Assert(i == e.Count() - 1);
-
-                        sb.AppendFormat(CultureInfo.InvariantCulture, 
-                            " WHEN AgeMin >= {0} THEN {1}", 
-                            d.MinimumAge, d.MinimumAge);
-                    }
-                }
-
-                sb.Append(" END");
-            }
-            else
-            {
-                sb.AppendFormat(CultureInfo.InvariantCulture,
-                    "UPDATE [{0}] SET AgeClass = {1}", 
-                    dataSheet.Name, e.ElementAtOrDefault(0).MinimumAge);
-            }
-
-            sb.AppendFormat(CultureInfo.InvariantCulture, " WHERE ScenarioID = {0}", dataSheet.Scenario.Id);
-            store.ExecuteNonQuery(sb.ToString());
-        }
-
-        public static DataTable CreateProportionChartData(Scenario scenario, ChartDescriptor descriptor, string tableName, DataStore store)
+        public static DataTable CreateProportionChartData(
+            Scenario scenario, 
+            ChartDescriptor descriptor, 
+            string tableName, 
+            DataStore store)
         {
             Dictionary<string, double> dict = CreateAmountDictionary(scenario, descriptor, store);
 
@@ -221,7 +118,10 @@ namespace SyncroSim.STSim
             return dt;
         }
 
-        public static Dictionary<string, double> CreateAmountDictionary(Scenario scenario, ChartDescriptor descriptor, DataStore store)
+        public static Dictionary<string, double> CreateAmountDictionary(
+            Scenario scenario, 
+            ChartDescriptor descriptor, 
+            DataStore store)
         {
             string tag = GetChartCacheTag(descriptor);
             Dictionary<string, double> dict = new Dictionary<string, double>();
@@ -309,7 +209,12 @@ namespace SyncroSim.STSim
             return query;
         }
 
-        private static string CreateRawAttributeDataQuery(Core.Scenario scenario, ChartDescriptor descriptor, string tableName, string attributeTypeColumnName, int attributeTypeId)
+        private static string CreateRawAttributeDataQuery(
+            Scenario scenario, 
+            ChartDescriptor descriptor, 
+            string tableName, 
+            string attributeTypeColumnName, 
+            int attributeTypeId)
         {
             string ScenarioClause = string.Format(CultureInfo.InvariantCulture,
                 "([{0}]={1})", 
@@ -340,7 +245,7 @@ namespace SyncroSim.STSim
             return query;
         }
 
-        private static string CreateRawExternalVariableDataQuery(Core.Scenario scenario, ChartDescriptor descriptor)
+        private static string CreateRawExternalVariableDataQuery(Scenario scenario, ChartDescriptor descriptor)
         {
             string[] s = descriptor.VariableName.Split('-');
             Debug.Assert(s.Count() == 2 && s[0] == "stsim_ExternalVariable");
@@ -466,6 +371,45 @@ namespace SyncroSim.STSim
             return true;
         }
 
+        /// <summary>
+        /// Determinies if a descriptor in the specified collection has an age reference
+        /// </summary>
+        /// <param name="descriptors"></param>
+        /// <returns></returns>
+        /// <remarks></remarks>
+        public static bool HasAgeReference(ChartDescriptorCollection descriptors)
+        {
+            foreach (ChartDescriptor d in descriptors)
+            {
+                if (HasAgeReference(d))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// Determines if the specified descriptor has an age reference
+        /// </summary>
+        /// <param name="descriptor"></param>
+        /// <returns></returns>
+        public static bool HasAgeReference(ChartDescriptor descriptor)
+        {
+            if (descriptor.IncludeDataFilter != null && descriptor.IncludeDataFilter.Contains("AgeClass"))
+            {
+                return true;
+            }
+
+            if (descriptor.DisaggregateFilter != null && descriptor.DisaggregateFilter.Contains("AgeClass"))
+            {
+                return true;
+            }
+
+            return false;
+        }
+
         private static string GetChartCacheTag(ChartDescriptor descriptor)
         {
             if (HasAgeReference(descriptor))
@@ -476,6 +420,298 @@ namespace SyncroSim.STSim
             {
                 return null;
             }
+        }
+
+        public static void DeleteAgeRelatedCacheEntries(Scenario scenario)
+        {
+            string CacheFolder = StochasticTime.ChartCache.GetCacheFolderName(scenario);
+
+            foreach (string f in Directory.GetFiles(CacheFolder))
+            {
+                if (f.EndsWith(Constants.AGE_QUERY_CACHE_TAG, StringComparison.Ordinal))
+                {
+                    File.Delete(f);
+                }
+            }
+        }
+
+        public static void SetAgeClassUpdateTag(Project project)
+        {
+            if (!project.Tags.Contains(Constants.AGECLASS_UPDATE_REQUIRED_TAG))
+            {
+                project.Tags.Add(new Tag(Constants.AGECLASS_UPDATE_REQUIRED_TAG, null));
+            }
+        }
+
+        public static void ClearAgeClassUpdateTag(Project project)
+        {
+            if (project.Tags.Contains(Constants.AGECLASS_UPDATE_REQUIRED_TAG))
+            {
+                project.Tags.Remove(Constants.AGECLASS_UPDATE_REQUIRED_TAG);
+            }
+        }
+
+        public static bool HasAgeClassUpdateTag(Project project)
+        {
+            return project.Tags.Contains(Constants.AGECLASS_UPDATE_REQUIRED_TAG);
+        }
+
+        /// <summary>
+        /// Gets a collection of current age descriptors
+        /// </summary>
+        /// <param name="project"></param>
+        /// <returns></returns>
+        /// <remarks></remarks>
+        public static IEnumerable<AgeDescriptor> GetAgeDescriptors(Project project)
+        {
+            IEnumerable<AgeDescriptor> e = GetAgeGroupDescriptors(project);
+
+            if (e == null)
+            {
+                e = GetAgeTypeDescriptors(project);
+            }
+
+#if DEBUG
+            if (e != null)
+            {
+                Debug.Assert(e.Count() > 0);
+            }
+#endif
+
+            if (e == null)
+            {
+                return null;
+            }
+
+            if (e.Count() == 0)
+            {
+                return null;
+            }
+
+            return e;
+        }
+
+        /// <summary>
+        /// Gets the AgeDescriptor in position [max-1]
+        /// </summary>
+        /// <param name="project"></param>
+        /// <returns></returns>
+        public static AgeDescriptor GetNextToLastAgeDescriptor(Project project)
+        {
+            IEnumerable<AgeDescriptor> e = GetAgeTypeDescriptors(project);
+
+#if DEBUG
+            if (e != null)
+            {
+                Debug.Assert(e.Count() > 0);
+            }
+#endif
+
+            if (e == null)
+            {
+                return null;
+            }
+
+            if (e.Count() == 0)
+            {
+                return null;
+            }
+
+            if (e.Count() == 1)
+            {
+                return e.First();
+            }
+            else
+            {
+                List<AgeDescriptor> l = e.ToList();
+                return l[e.Count() - 2];
+            }
+        }
+
+        /// <summary>
+        /// Gets an enumeration of age descriptors from the age group data sheet
+        /// </summary>
+        /// <returns></returns>
+        /// <remarks></remarks>
+        public static IEnumerable<AgeDescriptor> GetAgeGroupDescriptors(Project project)
+        {
+            DataTable dt = project.GetDataSheet(Strings.DATASHEET_AGE_GROUP_NAME).GetData();
+            DataView dv = new DataView(dt, null, null, DataViewRowState.CurrentRows);
+
+            if (dv.Count == 0)
+            {
+                return null;
+            }
+
+            List<AgeDescriptor> lst = new List<AgeDescriptor>();
+            Dictionary<int, bool> dict = new Dictionary<int, bool>();
+
+            foreach (DataRowView drv in dv)
+            {
+                int value = Convert.ToInt32(drv[
+                    Strings.DATASHEET_AGE_GROUP_MAXIMUM_COLUMN_NAME],
+                    CultureInfo.InvariantCulture);
+
+                if (!dict.ContainsKey(value))
+                {
+                    lst.Add(new AgeDescriptor(value, value));
+                    dict.Add(value, true);
+                }
+            }
+
+            lst.Sort((AgeDescriptor ad1, AgeDescriptor ad2) =>
+            {
+                return ad1.MinimumAge.CompareTo(ad2.MinimumAge);
+            });
+
+            int Prev = 0;
+
+            foreach (AgeDescriptor ad in lst)
+            {
+                int t = ad.MinimumAge;
+
+                ad.MinimumAge = Prev;
+                Prev = t + 1;
+            }
+
+            lst.Add(new AgeDescriptor(Prev, null));
+
+#if DEBUG
+            Debug.Assert(lst.Count > 0);
+
+            foreach (AgeDescriptor ad in lst)
+            {
+                if (ad.MaximumAge.HasValue)
+                {
+                    Debug.Assert(ad.MinimumAge <= ad.MaximumAge.Value);
+                }
+            }
+#endif
+
+            lst[lst.Count - 1].MaximumAge = null;
+            return lst;
+        }
+
+        /// <summary>
+        /// Gets an enumeration of age descriptors from the age type data sheet
+        /// </summary>
+        /// <returns></returns>
+        /// <remarks></remarks>
+        public static IEnumerable<AgeDescriptor> GetAgeTypeDescriptors(Project project)
+        {
+            DataRow dr = project.GetDataSheet(Strings.DATASHEET_AGE_TYPE_NAME).GetDataRow();
+
+            if (dr != null)
+            {
+                if (dr[Strings.DATASHEET_AGE_TYPE_FREQUENCY_COLUMN_NAME] != DBNull.Value &&
+                    dr[Strings.DATASHEET_AGE_TYPE_MAXIMUM_COLUMN_NAME] != DBNull.Value)
+                {
+                    int f = Convert.ToInt32(dr[Strings.DATASHEET_AGE_TYPE_FREQUENCY_COLUMN_NAME], CultureInfo.InvariantCulture);
+                    int m = Convert.ToInt32(dr[Strings.DATASHEET_AGE_TYPE_MAXIMUM_COLUMN_NAME], CultureInfo.InvariantCulture);
+
+                    if (f <= m)
+                    {
+                        AgeHelper h = new AgeHelper(true, f, m);
+                        return h.GetAges();
+                    }
+                }
+            }
+
+            return null;
+        }
+
+        public static void UpdateAgeClassIfRequired(DataStore store, Project project)
+        {
+            if (project.Tags.Contains(Constants.AGECLASS_UPDATE_REQUIRED_TAG))
+            {
+                foreach (Scenario s in project.Library.Scenarios)
+                {
+                    if (!s.IsDeleted && s.IsResult && s.Project == project)
+                    {
+                        UpdateAgeClassWork(store, s);
+                        DeleteAgeRelatedCacheEntries(s);
+                    }
+                }
+            }
+        }
+
+        public static void UpdateAgeClassWork(DataStore store, Scenario s)
+        {
+            ChartingUtilities.UpdateAgeClassColumn(store, s.GetDataSheet(Strings.DATASHEET_OUTPUT_STRATUM_TRANSITION_NAME));
+            ChartingUtilities.UpdateAgeClassColumn(store, s.GetDataSheet(Strings.DATASHEET_OUTPUT_STRATUM_STATE_NAME));
+            ChartingUtilities.UpdateAgeClassColumn(store, s.GetDataSheet(Strings.DATASHEET_OUTPUT_STATE_ATTRIBUTE_NAME));
+            ChartingUtilities.UpdateAgeClassColumn(store, s.GetDataSheet(Strings.DATASHEET_OUTPUT_TRANSITION_ATTRIBUTE_NAME));
+        }
+
+        /// <summary>
+        /// Updates the age class column for the specified table
+        /// </summary>
+        /// <param name="store"></param>
+        /// <param name="dataSheet"></param>
+        /// <remarks></remarks>
+        public static void UpdateAgeClassColumn(DataStore store, DataSheet dataSheet)
+        {
+            IEnumerable<AgeDescriptor> e = GetAgeDescriptors(dataSheet.Project);
+
+            if (e == null)
+            {
+                return;
+            }
+
+            if (!AnyDataExists(store, dataSheet))
+            {
+                return;
+            }
+
+            if (!AgeDataExists(store, dataSheet))
+            {
+                return;
+            }
+
+            StringBuilder sb = new StringBuilder();
+            Debug.Assert(!(e.ElementAtOrDefault(e.Count() - 1).MaximumAge.HasValue));
+
+            if (e.Count() > 1)
+            {
+                sb.AppendFormat(CultureInfo.InvariantCulture,
+                    "UPDATE [{0}] SET AgeClass = CASE",
+                    dataSheet.Name);
+
+                sb.Append(" WHEN 0 THEN 0");
+
+                for (int i = 0; i < e.Count(); i++)
+                {
+                    AgeDescriptor d = e.ElementAtOrDefault(i);
+
+                    if (d.MaximumAge.HasValue)
+                    {
+                        Debug.Assert(i < e.Count() - 1);
+
+                        sb.AppendFormat(CultureInfo.InvariantCulture,
+                            " WHEN AgeMin >= {0} And AgeMax <= {1} THEN {2}",
+                            d.MinimumAge, d.MaximumAge.Value, d.MinimumAge);
+                    }
+                    else
+                    {
+                        Debug.Assert(i == e.Count() - 1);
+
+                        sb.AppendFormat(CultureInfo.InvariantCulture,
+                            " WHEN AgeMin >= {0} THEN {1}",
+                            d.MinimumAge, d.MinimumAge);
+                    }
+                }
+
+                sb.Append(" END");
+            }
+            else
+            {
+                sb.AppendFormat(CultureInfo.InvariantCulture,
+                    "UPDATE [{0}] SET AgeClass = {1}",
+                    dataSheet.Name, e.ElementAtOrDefault(0).MinimumAge);
+            }
+
+            sb.AppendFormat(CultureInfo.InvariantCulture, " WHERE ScenarioID = {0}", dataSheet.Scenario.Id);
+            store.ExecuteNonQuery(sb.ToString());
         }
     }
 }
